predictor now geenrate a full coco file
